refactor: CreateIndexesOperation upgraded to use CommandOperationV2
