+++ lib/collection.js
@@ -36,7 +36,7 @@ const AggregateOperation = require('./operations/aggregate');
 const BulkWriteOperation = require('./operations/bulk_write');
 const CountDocumentsOperation = require('./operations/count_documents');
 const CreateIndexOperation = require('./operations/create_index');
-const CreateIndexesOperation = require('./operations/create_indexes');
+const CreateIndexesOperation = require('./operations/create_indexes_v2');
 const DeleteManyOperation = require('./operations/delete_many');
 const DeleteOneOperation = require('./operations/delete_one');
 const DistinctOperation = require('./operations/distinct');
