+++ lib/operations/create_index.js
@@ -6,7 +6,6 @@ const CommandOperationV2 = require('./command_v2');
 const parseIndexOptions = require('../utils').parseIndexOptions;
 const MongoError = require('../error').MongoError;
 const maxWireVersion = require('../core/utils').maxWireVersion;
-const kCommand = Symbol('createIndexCommand');
 
 const keysToOmit = new Set([
   'name',
@@ -53,15 +52,7 @@ class CreateIndexOperation extends CommandOperationV2 {
         indexes[optionName] = options[optionName];
       }
     }
-
-    // build the command
-    const cmd = { createIndexes: this.name, indexes: [indexes] };
-
-    if (options.commitQuorum != null) {
-      cmd.commitQuorum = options.commitQuorum;
-    }
-
-    this[kCommand] = cmd;
+    this.indexes = indexes;
   }
 
   /**
@@ -79,14 +70,17 @@ class CreateIndexOperation extends CommandOperationV2 {
       });
     }
 
-    const cmd = this[kCommand];
+    const cmd = { createIndexes: this.name, indexes: [this.indexes] };
 
     // Ensure commitQuorum not passed if server < 4.4
-    if (cmd.commitQuorum != null && maxWireVersion(server) < 9) {
-      throw MongoError.create({
-        message: '`commitQuorum` option for `createIndexes` not supported on servers < 4.4',
-        driver: true
-      });
+    if (this.options.commitQuorum != null) {
+      if (maxWireVersion(server) < 9) {
+        throw MongoError.create({
+          message: '`commitQuorum` option for `createIndexes` not supported on servers < 4.4',
+          driver: true
+        });
+      }
+      cmd.commitQuorum = this.options.commitQuorum;
     }
 
     super.executeCommand(server, cmd, (err, result) => {
