+++ lib/operations/create_indexes.js
@@ -2,26 +2,42 @@
 
 const Aspect = require('./operation').Aspect;
 const defineAspects = require('./operation').defineAspects;
-const OperationBase = require('./operation').OperationBase;
-const executeCommand = require('./db_ops').executeCommand;
-const MongoError = require('../core').MongoError;
-const ReadPreference = require('../core').ReadPreference;
+const CommandOperationV2 = require('./command_v2');
+const MongoError = require('../error').MongoError;
 const maxWireVersion = require('../core/utils').maxWireVersion;
+const ReadPreference = require('../core').ReadPreference;
 
-class CreateIndexesOperation extends OperationBase {
+/**
+ * Creates one or more indexes on the collection.
+ *
+ * @class
+ */
+class CreateIndexesOperation extends CommandOperationV2 {
+  /**
+   * Construct a CreateIndexes operation.
+   *
+   * @param {Collection} collection parent collection
+   * @param {string|object} indexSpecs Defines the indexes.
+   * @param {object} [options] Optional settings. See Collection.prototype.createIndexes for a list of options.
+   */
   constructor(collection, indexSpecs, options) {
-    super(options);
-
+    super(collection, options);
     this.collection = collection;
     this.indexSpecs = indexSpecs;
   }
 
-  execute(callback) {
+  /**
+   * Execute the operation.
+   *
+   * @param {any} server
+   * @param {Collection~resultCallback} [callback] The command result callback
+   */
+  execute(server, callback) {
     const coll = this.collection;
+    const options = Object.assign({}, this.options, { readPreference: ReadPreference.PRIMARY });
     const indexSpecs = this.indexSpecs;
-    let options = this.options;
 
-    const capabilities = coll.s.topology.capabilities();
+    const collationNotSupported = maxWireVersion(server) < 5;
 
     // Ensure we generate the correct name if the parameter is not set
     for (let i = 0; i < indexSpecs.length; i++) {
@@ -29,7 +45,7 @@ class CreateIndexesOperation extends OperationBase {
         const keys = [];
 
         // Did the user pass in a collation, check if our write server supports it
-        if (indexSpecs[i].collation && capabilities && !capabilities.commandsTakeCollation) {
+        if (indexSpecs[i].collation && collationNotSupported) {
           return callback(new MongoError('server/primary/mongos does not support collation'));
         }
 
@@ -42,9 +58,7 @@ class CreateIndexesOperation extends OperationBase {
       }
     }
 
-    options = Object.assign({}, options, { readPreference: ReadPreference.PRIMARY });
-
-    const command = {
+    const cmd = {
       createIndexes: coll.collectionName,
       indexes: indexSpecs
     };
@@ -59,14 +73,20 @@ class CreateIndexesOperation extends OperationBase {
           })
         );
       }
-      command.commitQuorum = options.commitQuorum;
+      cmd.commitQuorum = options.commitQuorum;
     }
 
-    // Execute the index
-    executeCommand(coll.s.db, command, options, callback);
+    super.executeCommand(server, cmd, (err, result) => {
+      if (err) {
+        callback(err);
+        return;
+      }
+
+      callback(null, this.options.full ? result : this.indexName);
+    });
   }
 }
 
-defineAspects(CreateIndexesOperation, Aspect.WRITE_OPERATION);
+defineAspects(CreateIndexesOperation, [Aspect.WRITE_OPERATION, Aspect.EXECUTE_WITH_SELECTION]);
 
 module.exports = CreateIndexesOperation;
